add connection observer

+++ app/src/main/java/com/example/warmmeal/fragment_home/view/HomeFragment.java
@@ -24,6 +24,7 @@ import com.example.warmmeal.fragment_favourite.view.OnDeleteFromFavouriteRespons
 import com.example.warmmeal.fragment_favourite.view.OnGetFavouriteMealResponse;
 import com.example.warmmeal.fragment_home.presenter.HomeFragmentPresenter;
 import com.example.warmmeal.fragment_home.view.adapters.HomeRecyclerViewAdapter;
+import com.example.warmmeal.fragment_home.view.contracts.IHomeFragment;
 import com.example.warmmeal.fragment_home.view.contracts.OnNestedRecyclerViewItemClickedListener;
 import com.example.warmmeal.fragment_home.view.contracts.OnNetworkCallResponse;
 import com.example.warmmeal.meal_screen.view.MealActivity;
@@ -44,7 +45,11 @@ import com.google.android.material.snackbar.Snackbar;
 import java.util.ArrayList;
 import java.util.List;
 
-public class HomeFragment extends Fragment implements OnNestedRecyclerViewItemClickedListener, OnNetworkCallResponse, OnAddToFavouriteResponse, OnGetFavouriteMealResponse , OnDeleteFromFavouriteResponse {
+import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers;
+import io.reactivex.rxjava3.core.Flowable;
+import io.reactivex.rxjava3.schedulers.Schedulers;
+
+public class HomeFragment extends Fragment implements OnNestedRecyclerViewItemClickedListener, OnNetworkCallResponse, OnAddToFavouriteResponse, OnGetFavouriteMealResponse , OnDeleteFromFavouriteResponse , IHomeFragment {
 
 
 
@@ -109,9 +114,11 @@ public class HomeFragment extends Fragment implements OnNestedRecyclerViewItemCl
         favouriteMeals = new ArrayList<>();
         context = view.getContext();
 
-        presenter = HomeFragmentPresenter.getInstance(RepositoryImpl.getInstance(FirebaseHandler.getInstance(), NetworkAPI.getInstance(), DatabaseHandler.getInstance(context), SharedPrefHandler.getInstance(context)));
+        presenter = HomeFragmentPresenter.getInstance(RepositoryImpl.getInstance(FirebaseHandler.getInstance(), NetworkAPI.getInstance(), DatabaseHandler.getInstance(context), SharedPrefHandler.getInstance(context)),this);
 
-        presenter.getMealsByFirstLetter('a',DataPurpose.INSPIRATION,this);
+        //presenter.getMealsByFirstLetter('a',DataPurpose.INSPIRATION,this);
+        presenter.getRandomMeal();
+        presenter.getMealsByFirstLetter('b',DataPurpose.MORE_YOU_LIKE,this);
         presenter.getAllCategories(this);
         presenter.getAllCountries(this);
 
@@ -124,11 +131,7 @@ public class HomeFragment extends Fragment implements OnNestedRecyclerViewItemCl
 
     }
 
-    @Override
-    public void onStart() {
-        super.onStart();
-        presenter.getMealsByFirstLetter('b',DataPurpose.MORE_YOU_LIKE,this);
-    }
+
 
     void setUpRecyclerViewWithLists()
     {
@@ -162,12 +165,14 @@ public class HomeFragment extends Fragment implements OnNestedRecyclerViewItemCl
 
     @Override
     public void onAddToFavouriteClicked(Meal meal) {
-        if (!meal.isFavourite())
+        if (!meal.isFavourite() && FirebaseHandler.CURRENT_USER_ID != null)
         {
             presenter.deleteFromFavourite(new FavouriteMeal(FirebaseHandler.CURRENT_USER_ID,meal.getIdMeal(),meal.getStrMeal(),meal.getStrMealThumb(),true),this);
-        }else
+        }else if(FirebaseHandler.CURRENT_USER_ID != null)
         {
             presenter.addFavouriteMeal(new FavouriteMeal(FirebaseHandler.CURRENT_USER_ID,meal.getIdMeal(),meal.getStrMeal(),meal.getStrMealThumb(),true),this);
+        }else {
+            Toast.makeText(context, "Please login first", Toast.LENGTH_SHORT).show();
         }
     }
 
@@ -188,6 +193,7 @@ public class HomeFragment extends Fragment implements OnNestedRecyclerViewItemCl
     @Override
     public void onGetMealByCharacterForMoreYouLikeSuccess(Meals meals) {
         mealsYouMightLike = (ArrayList<Meal>) meals.getMeals();
+        setUpRecyclerViewWithLists();
         presenter.getAllFavouriteMeals(FirebaseHandler.CURRENT_USER_ID,this);
 
     }
@@ -214,10 +220,22 @@ public class HomeFragment extends Fragment implements OnNestedRecyclerViewItemCl
     }
 
     @Override
+    public void onGetRandomMealSuccess(Meals meals) {
+        dailyInspirationMeals.add(meals.getMeals().get(0));
+    }
+
+    @Override
+    public void onGetRandomMealComplete() {
+        setUpRecyclerViewWithLists();
+    }
+
+    @Override
     public void onFailure(String message) {
         Log.d("Kerolos", "Home Fragment onFailure: " + message);
         customProgressBar.dismissProgressBar();
     }
+
+
 
     ////////
 
@@ -242,14 +260,16 @@ public class HomeFragment extends Fragment implements OnNestedRecyclerViewItemCl
 
     @Override
     public void onGetFavouriteMealSuccess(List<FavouriteMeal> favouriteMeals) {
-        if(!isFavouriteMealsFetched)
+        if(!isFavouriteMealsFetched && FirebaseHandler.CURRENT_USER_ID != null)
         {
-            Log.d("Kerolos", "onGetFavouriteMealSuccess: " + favouriteMeals.size());
             FavouriteMeal.getFavouriteMealsList((ArrayList<FavouriteMeal>) favouriteMeals, dailyInspirationMeals);
             FavouriteMeal.getFavouriteMealsList((ArrayList<FavouriteMeal>) favouriteMeals, mealsYouMightLike);
             setUpRecyclerViewWithLists();
             customProgressBar.dismissProgressBar();
             isFavouriteMealsFetched = true;
+        }
+        else {
+            customProgressBar.dismissProgressBar();
         }
     }
